third_party/blink/web_tests/http/tests/xmlhttprequest/post-formdata-expected.txt

Test verifies that FormData is sent correctly when using XMLHttpRequest asynchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS echoResult is "string=string value"
PASS echoResult is "bareBlob=blob:application/octet-stream:blob-value"
PASS echoResult is "mimeBlob=blob:text/html:blob-value"
PASS echoResult is "namedBlob=blob-file.txt:application/octet-stream:blob-value"
PASS echoResult is "bareFile=file-name.txt:application/octet-stream:file-value"
PASS echoResult is "mimeFile=file-name.html:text/html:file-value"
PASS echoResult is "renamedFile=file-name-override.html:text/html:file-value"
PASS successfullyParsed is true

TEST COMPLETE
